+++ src/backend/config.rs
@@ -14,13 +14,13 @@ const APP_TITLE_DEFAULT: &str = "TYT";
 const UPDATAE_AT_START_DEFAULT: bool = true;
 const SORT_BY_TAG_DEFAULT: bool = false;
 const MASSAGE_TIMEOUT_DEFAULT: usize = 20;
-const USE_NOTIFY_SEND_DEFAULT: bool = true;
+const NOTIFY_WITH_DEFAULT: &str = "notify_send";
 const VIDEO_PLAYER_DEFAULT: &str = "mpv";
 const DEFAULT_SORT: SortingMethod = SortingMethod::Date;
 
 #[derive(Clone, Deserialize, Serialize)]
 #[serde(default)]
-pub struct Config {
+pub(crate) struct Config {
     pub show_empty_channels: bool,
     pub mark_on_open: bool,
     pub down_on_mark: bool,
@@ -28,9 +28,9 @@ pub struct Config {
     pub update_at_start: bool,
     pub sort_by_tag: bool,
     pub message_timeout: usize,
-    pub use_notify_send: bool,
     pub video_player: String,
     pub default_sorting_method: SortingMethod,
+    pub notify_with: String,
 }
 
 impl Default for Config {
@@ -43,7 +43,7 @@ impl Default for Config {
             update_at_start: UPDATAE_AT_START_DEFAULT,
             sort_by_tag: SORT_BY_TAG_DEFAULT,
             message_timeout: MASSAGE_TIMEOUT_DEFAULT,
-            use_notify_send: USE_NOTIFY_SEND_DEFAULT,
+            notify_with: NOTIFY_WITH_DEFAULT.into(),
             video_player: VIDEO_PLAYER_DEFAULT.into(),
             default_sorting_method: DEFAULT_SORT,
         }
@@ -51,7 +51,7 @@ impl Default for Config {
 }
 
 impl Config {
-    pub fn read_config_file() -> Self {
+    pub(crate) fn init() -> Self {
         let mut path = home_dir().unwrap();
         path.push(CONFIG_FILE_PATH);
 
@@ -73,7 +73,7 @@ impl Config {
                         let string = serde_yaml::to_string(&def_config).unwrap();
 
                         match file.write_all(string.as_bytes()) {
-                            Ok(_) => Config::read_config_file(),
+                            Ok(_) => Config::init(),
                             Err(e) => panic!("could not write default config: {}", e),
                         }
                     }
